pyproject.toml
[tool.poetry]
name = "DGE-ANN"
version = "0.1.0"
description = "Node Attention Networks in Item Retrieval Task"
authors = ["Androsov Dmytro"]

[tool.poetry.dependencies]
python = ">=3.9.2 <3.10"
dvc = "^2.10.0"
hydra-core = "^1.1.1"
pdoc3 = "^0.10.0"
notebook = "^6.4.10"
bazel = "^0.0.0"
tensorflow-macos = "^2.8.0"
tensorflow-recommenders = "^0.3.0"
networkx = "^2.8"
graphviz = "^0.20"
pandas = "^1.4.2"
numpy = "^1.22.3"
scipy = "^1.8.0"
tensorflow-addons = "^0.16.1"
tensorflow-ranking = "^0.1.5"
tqdm = "^4.64.0"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build   
)/ 
'''
